rag-system/chatbot-history/requirements.txt
# === Core dependencies (đã có) ===
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.120.0
filelock==3.20.0
fsspec==2025.9.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openai==2.6.1
packaging==25.0
pillow==12.0.0
portalocker==2.10.1
protobuf==6.33.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.0.0
# pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.10.23
requests==2.31.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==3.0.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1

# === Data ingestion dependencies ===
beautifulsoup4==4.12.3
feedparser==6.0.11
wikipedia-api==0.8.1
schedule==1.2.2

apscheduler==3.10.4
pytz==2023.3

# PostgreSQL & SQLAlchemy
sqlalchemy==2.0.36
psycopg[binary]==3.2.5
alembic==1.13.3

perplexityai